requirements.txt
discord.py>=2.3.0
wavelink>=3.0.0
python-dotenv
aiosqlite